Prepare to consolidate route helpers
